Simplified fixDisplacedIndex for the new adds
It used to search around and stuff in case the new
ones were out of order. Now we trust they're in order
and add left or right. So, nothing is necessary on a normal
add, and on an add:left (triggered by _addLeft), we just
increment TractRing.currentTractIndex by one.