Implemented setRenderDistance and associated fns
TractViews can now be told what render distance they are required
to maintain.  setRenderDistance also renders them.

When they are rendered, they will create imgDivs for all images that
are in range and don't yet have imgDivs.

Right now, on init, they do this.setRenderDistance(true, 0).  This has
identical behavior to the way it used to work - everything is on and keeps
its old imgDivs, but they are loaded only when each img is viewed.

Next, need to write the code that sets the render distances on various
views to appropriate values.