Rudimentary img caching scheme added to manager
The TractDataManager now responds to 'refresh' and
'tract:nav' events on the ring by setting the context
for the current tract to 3, the adjacent tracts to 1 and
the doubly adjacent tracts to 0.

Views initialize themselves to (true, 0) because initializing
to (false, 0) fails with an error related to a line near 473,
since view.render() is undefined when view.render() is called
on a view with its on/off set to off. I believe this is just
because it returns nothing, rather than returning this to permit
chaining (which it does when it renders normally.)  This will be
fixed next.