Added educational attainment transform and transforms_columns.py
Transforms columns is just a new place to keep the tuples of column
names for each transformation. It is imported into the data
processor as cols, so cols.age_distribution is the tuple of column
names for the age distributions. Seems cleaner than having all
those huge tuples in the actual processor code.

Educational attainment is given for people ages 18+. The distribution
is a list of number of people with attainments of...

<9th grade
Some HS, no degree
HS degree/GED/equivalent
Some college, no degree
Associates Degree
Bachelors Degree
Doctoral or Professional Degree

Educational attainment also comes with a 'total' subcategory, which
is just the total number of people >18 years of age, so that the
client can calculate percentages, which it will presumably want.

This leads to a question of when we should provide the client with
raw population counts and when we should just provide percentages.
For some reason I did % for race and raw # for this.