Added pct_white_not_latino stat and pctile of it
Since before I did it off raw numbers of white not latino
people and that was stupid.

Also changed quintilebg in censusapp.js to try{}catch(e){}
and return the neutral value if the argument was no good
(not a number between 0.0 and 1.0) so that that checking doesn't
have to be repeated in every display fn.