Experiment: "highest" on diversity is low %white
The lowest-highest thing seems to make most sense
as "lowest diversity" (highest % white people who
are not hispanic) or "highest diversity" (lowest %
white people who are not hispanic), but this goes
against a general system of big numbers == highest.

Not sure. The stat %non-white is more awkward to think
about though I think.